requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
jinja2==3.1.2
weasyprint==60.1
grpcio-reflection==1.59.3